move api under barcode controller
