+++ src/main/java/it/gov/pagopa/payment/controller/payment/BarCodePaymentController.java
@@ -6,6 +6,7 @@ import it.gov.pagopa.payment.dto.PreviewPaymentRequestDTO;
 import it.gov.pagopa.payment.dto.barcode.AuthBarCodePaymentDTO;
 import it.gov.pagopa.payment.dto.barcode.TransactionBarCodeCreationRequest;
 import it.gov.pagopa.payment.dto.barcode.TransactionBarCodeResponse;
+import it.gov.pagopa.payment.dto.qrcode.TransactionResponse;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
@@ -43,5 +44,9 @@ public interface BarCodePaymentController {
             @PathVariable("initiativeId") String initiativeId,
             @RequestHeader("x-user-id") String userId
     );
+
+    @PutMapping("/bar-code/{trxCode}/capture")
+    TransactionResponse capturePayment(
+            @PathVariable("trxCode") String trxCode);
 
 }
