+++ src/main/java/it/gov/pagopa/payment/controller/payment/CommonPaymentControllerImpl.java
@@ -15,20 +15,17 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class CommonPaymentControllerImpl implements CommonPaymentController {
     private final CommonCreationServiceImpl commonCreationService;
-    private final CommonCaptureServiceImpl commonCaptureService;
     private final CommonConfirmServiceImpl commonConfirmService;
     private final CommonCancelServiceImpl commonCancelService;
     private final CommonStatusTransactionServiceImpl commonStatusTransactionService;
     private final QRCodeExpirationService qrCodeExpirationService; // used just to force the expiration: this behavior is the same for all channels
 
     public CommonPaymentControllerImpl(@Qualifier("commonCreate") CommonCreationServiceImpl commonCreationService,
-                                       @Qualifier("commonCapture") CommonCaptureServiceImpl commonCaptureService,
                                        @Qualifier("commonConfirm") CommonConfirmServiceImpl commonConfirmService,
                                        @Qualifier("commonCancel") CommonCancelServiceImpl commonCancelService,
                                        CommonStatusTransactionServiceImpl commonStatusTransactionService,
                                        QRCodeExpirationService qrCodeExpirationService) {
         this.commonCreationService = commonCreationService;
-        this.commonCaptureService= commonCaptureService;
         this.commonConfirmService = commonConfirmService;
         this.commonCancelService = commonCancelService;
         this.commonStatusTransactionService = commonStatusTransactionService;
@@ -59,14 +56,6 @@ public class CommonPaymentControllerImpl implements CommonPaymentController {
                 acquirerId,
                 trxId);
         return commonConfirmService.confirmPayment(trxId, merchantId, acquirerId);
-    }
-
-    @Override
-    @PerformanceLog(
-            value = "CAPTURE_PAYMENT",
-            payloadBuilderBeanClass = TransactionResponsePerfLoggerPayloadBuilder.class)
-    public TransactionResponse capturePayment(String trxCode) {
-        return commonCaptureService.capturePayment(trxCode);
     }
 
     @Override
